gcp/requirements.txt
tensorflow==2.9.0
google-cloud-storage==1.16.1
pillow